streamlit/requirements.txt
streamlit>=1.35.0
streamlit-folium>=0.20.0
streamlit-autorefresh>=1.0.1
pandas>=2.0.0
plotly>=5.18.0
folium>=0.16.0
psycopg2-binary>=2.9.9
openpyxl>=3.1.2
rapidfuzz>=3.6.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
